Dangling open lock files cause Windows test failures
Log manager code would fail to close an open lock file if it was
unsuccessful in actually locking the file.  On Windows this would lead
to automated test failures as cleanup code attempted to remove the
directory while the lock file remained open.  This could be seen
for example in:

LogManagerTest.testTwoLogManagersUsingSameDirFails