KAFKA-5437; Always send a sig_kill when cleaning the message copier
When the message copier hangs (like when there is a bug in the client), it ignores the sigterm and doesn't shut down. this leaves the cluster in an unclean state causing future tests to fail.

In this patch we always send SIGKILL when cleaning the node if the process isn't already dead. This is consistent with the other services.

Author: Apurva Mehta <apurva@confluent.io>

Reviewers: Jason Gustafson <jason@confluent.io>

Closes #3308 from apurvam/KAFKA-5437-force-kill-message-copier-on-cleanup