KAFKA-5438; Fix UnsupportedOperationException in WriteTxnMarkersRequest
Before this patch, the `partitionErrors` was an immutable map. As a result if a single producer had a marker for multiple partitions, and if there were multiple response callbacks for a single append, we would get an `UnsupportedOperationException` in the `writeTxnMarker` handler.

Author: Apurva Mehta <apurva@confluent.io>

Reviewers: Ismael Juma <ismael@juma.me.uk>, Jason Gustafson <jason@confluent.io>

Closes #3313 from apurvam/KAFKA-5438-fix-unsupportedoperationexception-in-writetxnmarker

(cherry picked from commit c2daf4c623e5919872012aff57ffa25af46a97f8)
Signed-off-by: Jason Gustafson <jason@confluent.io>