KAFKA-5472: Eliminated duplicate group names when validating connector results
Kafka Connect was adding duplicate group names in the response from the REST API's validation of connector configurations. This fixes the duplicates and maintains the order of the `ConfigDef` objects so that the `ConfigValue` results are in the same order.

This is a blocker and should be merged to 0.11.0.

Author: Randall Hauch <rhauch@gmail.com>

Reviewers: Ewen Cheslack-Postava <ewen@confluent.io>

Closes #3379 from rhauch/KAFKA-5472

(cherry picked from commit de982ba3fbf99664f0aaa5aa4b72af8fd1881232)
Signed-off-by: Ewen Cheslack-Postava <me@ewencp.org>