KAFKA-3575: Use console consumer access topic that does not exist, can not use "Control + C" to exit process
A finally block is not guaranteed to execute in the event of Ctrl+C happening
while in the try or catch blocks. Decrementing the latch in the finally block
therefore made the shutdown hook hang waiting for something that would
never happen and the JVM couldn't exit while it was running the shutdown hook.

Replacing the latch with an atomic flag to say whether we've run the cleanup
code allows us to either run it from the shutdown hook, or the finally block.
It should thus definitely run once. When run from the shutdown hook the main
thread would no longer be running, so it should be threadsafe.