KAFKA-5032; Update the docs for message size configs across the board
Before 0.11, we used to have limits for maximum message size on the producer, broker, and consumer side.

From 0.11 onward, these limits apply to record batches as a whole. This patch updates the documentation of the configs to make this explicit.

A separate patch will have more extensive upgrade notes to tie all the changes together in one narrative.

Author: Apurva Mehta <apurva@confluent.io>

Reviewers: Jason Gustafson <jason@confluent.io>, Ismael Juma <ismael@juma.me.uk>

Closes #3374 from apurvam/KAFKA-5032-message-size-docs