KAFKA-5630; Consumer should block on corrupt records and keep throwing an exception
This patch handles the case that a CorruptRecordException is thrown from the iterator directly.
The fix is a little tricky as exceptions can be thrown from a few different scenarios. The current approach is to let the same record go through the exact same process as last time when exception is thrown, so the exception will be thrown at the same step. The only problem for that is the iterator state will change once it throws an exception. To handle that we cache the first iterator exception and put it into the suppressed exception of the IllegalStateException thrown in the future.

Author: Jiangjie Qin <becket.qin@gmail.com>

Reviewers: Jason Gustafson <jason@confluent.io>

Closes #3573 from becketqin/KAFKA-5630

(cherry picked from commit cd4c6c0183edfbf8e7e8ce526d638e31fcd09186)
Signed-off-by: Jason Gustafson <jason@confluent.io>