KAFKA-5731; Corrected how the sink task worker updates the last committed offsets (0.11.0)
Prior to this change, it was possible for the synchronous consumer commit request to be handled before previously-submitted asynchronous commit requests. If that happened, the out-of-order handlers improperly set the last committed offsets, which then became inconsistent with the offsets the connector task is working with.

This change ensures that the last committed offsets are updated only for the most recent commit request, even if the consumer reorders the calls to the callbacks.

**This is for the `0.11.0` branch; see #3662 for the equivalent and already-approved PR for `trunk`.**

Author: Randall Hauch <rhauch@gmail.com>

Reviewers: Jason Gustafson <jason@confluent.io>

Closes #3672 from rhauch/kafka-5731-0.11.0