KAFKA-5659; Fix error handling, efficiency issue in AdminClient#describeConfigs
If a request for a broker configuration failed due to a timeout or
the broker not being available, we would fail the futures
associated with the non-broker request instead (and never fail the
broker future, which would be left uncompleted forever).

We would also do an unnecessary request if only broker configs were
requested.

Author: Colin P. Mccabe <cmccabe@confluent.io>

Reviewers: Ismael Juma <ismael@juma.me.uk>

Closes #3585 from cmccabe/KAFKA-5659