Handled remaining corner cases:
  1. Batch expiry on the producer. In this case, we need mute the
partition and make sure all the inflight requests are fully resolved. If
they are successful, then all is well. If they are also expired or fail
due to OutOfOrderSequence, we need to reset the producer id.

  2. We now deal correctly with out of order responses to in flight
requests which may happen due to leadership changes. We keep track of
all inflight batches, ordered by sequence. Whenever there is an error,
we reduce to one inflight request by ensuring only the next batch
by sequence can be drained. When we requeue a batch, we check to make
sure that the queue remains in sequence order. When a batch completes,
it is removed from the in flight list, and the next batch by sequence
changes.

Some other changes:

 1. We never change sequence numbers unless we are certain that a batch
failed fatally on the broker (got a fatal error code), _and_ future
batches retured OutOfOrderSequence. Before the latest commit, we would
always reset sequence numbers on retry, and have other mechanisms to
ensure that a given batch got the same sequence as before on the next
drain. This is now much tighter, and needs to be in order to correctly
handle the out of order response problem.

Todo:

Add unit tests for out of order responses.
Add unit tests new cases around batch expiry (all batches expired, all
batches succeeded, future batches failed with OutOfSequence).