Address PR comments.
One notable update: We didn't have a test for reassiging sequence
numbers, and we had a bug in that logic. Essentially a newly reassigned
batch would be fail with an out of order sequence exception because it
would be the next in line batch after its sequence was adjusted.

This is fixed by tracking whether the sequence of a batch has been
adjusted since the last drain, in which case the out of order sequence
exception for it is always retried. If it fails again, it will be fatal.