KAFKA-5899 Make metric names even more restrictive than KIP-196 specifies
KIP-196 specifies that the connector and task names are used as tags within the metric group names, and that these tag values will consist of the characters `[.-\w]+`. This matches the existing Kafka metric documentation (for client ID, etc.), but the `.` character can cause problems in Graphite where `.` signals a hierarchical name.

This commit changes the `makeValidName` utility to replace with a single `-` character all sequences of 1 or more `-` and word characters (e.g., `[a-zA-Z_0-9]`). Note that `.` will also be replaced. This is more restrictive than the KIP, and thus should be valid.