KAFKA-5902 Changed to measuring task processing lag behind consumer
Changed the `sink-record-lag-max` metric to be the maximum lag in terms of number of records that the sink task is behind the consumer's position for any topic partitions. This is not ideal, since often “lag” is defined to represent how far behind the task (or consumer) is relative to the end of the topic partition. However, the most recent offset for the topic partition is not easy to access in Connect.