KAFKA-5986; Streams State Restoration never completes when logging is disabled
When logging is disabled and there are state stores the task never transitions from restoring to running. This is because we only ever check if the task has state stores and return false on initialization if it does. The check should be if we have changelog partitions, i.e., we need to restore.

Author: Damian Guy <damian.guy@gmail.com>

Reviewers: Matthias J. Sax <matthias@confluent.io>, Bill Bejeck <bill@confluent.io>, tedyu <yuzhihong@gmail.com>, Ismael Juma <ismael@juma.me.uk>

Closes #3983 from dguy/restore-test

(cherry picked from commit 3107a6c5c8d1358b8e705c5d5a16b7441d2225a6)
Signed-off-by: Damian Guy <damian.guy@gmail.com>