KAFKA-6003; Accept appends on replicas unconditionally when local producer state doesn't exist
Without this patch, if the replica's log was somehow truncated before
the leader's, it is possible for the replica fetcher thread to
continuously throw an OutOfOrderSequenceException because the
incoming sequence would be non-zero and there is no local state.

This patch changes the behavior so that the replica state is updated to
the leader's state if there was no local state for the producer at the
time of the append.

Author: Apurva Mehta <apurva@confluent.io>

Reviewers: Ismael Juma <ismael@juma.me.uk>, Jason Gustafson <jason@confluent.io>

Closes #4004 from apurvam/KAFKA-6003-handle-unknown-producer-on-replica

(cherry picked from commit 6ea4fffdd287a0c6a02c1b6dc1006b1a7b614405)
Signed-off-by: Jason Gustafson <jason@confluent.io>